requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncio==4.0.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
distro==1.9.0
fastapi==0.136.0
filelock==3.28.0
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.11
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.15
langchain-core==1.3.0
langchain-openai==1.1.14
langgraph==1.1.8
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
langsmith==0.7.32
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.19.0
networkx==3.6.1
numpy==2.4.4
openai==2.32.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.1
pandas==3.0.2
pillow==12.2.0
plotly==6.7.0
protobuf==7.34.1
pyarrow==23.0.1
pydantic==2.13.2
pydantic_core==2.46.2
pydeck==0.9.2
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
starlette==1.0.0
streamlit==1.56.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
watchdog==6.0.0
xxhash==3.6.0
zstandard==0.25.0